Move errorPage.tsx
