+++ ui/src/appRoutes.tsx
@@ -4,11 +4,11 @@ import { AddCohortCriteria, AddFeatureSetCriteria } from "addCriteria";
 import { CohortReview } from "cohortReview/cohortReview";
 import { CohortReviewList } from "cohortReview/cohortReviewList";
 import CohortRoot from "cohortRoot";
+import { ErrorPage } from "components/errorPage";
 import { StudySourceContextRoot } from "data/studySourceContext";
 import { UnderlaySourceContextRoot } from "data/underlaySourceContext";
 import Edit from "edit";
 import { getEnvironment } from "environment";
-import { ErrorPage } from "errorhandler";
 import { Export } from "export";
 import { AddFeatureSet } from "featureSet/addFeatureSet";
 import { FeatureSet } from "featureSet/featureSet";
